naming fix

+++ src/main/kotlin/com/wafflestudio/toyproject/memoWithTags/user/controller/AdminController.kt
@@ -28,21 +28,21 @@ class AdminController(
     @Operation(summary = "유저 목록 가져오기")
     @GetMapping("/admin/user")
     fun getUsers(@AuthUser user: User): List<User> {
-        adminService.isAdmin(user.id)
+        adminService.assertAdminRole(user.id)
         return adminService.getUsers()
     }
 
     @Operation(summary = "유저 계정 생성")
     @PostMapping("/admin/user")
     fun createUser(@AuthUser user: User, @RequestBody request: CreateUserRequest): User {
-        adminService.isAdmin(user.id)
-        return userService.register(request.email, request.password)
+        adminService.assertAdminRole(user.id)
+        return userService.register(request.email, request.nickname, request.password)
     }
 
     @Operation(summary = "유저 계정 삭제하기")
     @DeleteMapping("/admin/user/{userId}")
     fun deleteUser(@AuthUser user: User, @PathVariable userId: UUID): ResponseEntity<Unit> {
-        adminService.isAdmin(user.id)
+        adminService.assertAdminRole(user.id)
         adminService.deleteUser(userId)
         return ResponseEntity.ok().build()
     }
@@ -50,7 +50,7 @@ class AdminController(
     @Operation(summary = "유저 계정 업데이트")
     @PutMapping("/admin/user")
     fun getUsers(@AuthUser user: User, @RequestBody request: UpdateUserRequest): User {
-        adminService.isAdmin(user.id)
+        adminService.assertAdminRole(user.id)
         return adminService.updateUser(request.id, request.userUpdateInfo)
     }
 }
